% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_offsets.R
\name{calculate_offsets}
\alias{calculate_offsets}
\title{Calculates the offsets for the current block}
\usage{
calculate_offsets(
  current_missings,
  current_observations,
  mcontrol,
  current_block,
  pred,
  liste,
  X,
  blocks,
  current_intercept
)
}
\arguments{
\item{current_missings}{index vector (indices) of current missing observations}

\item{current_observations}{index vector (indices) of current used observations}

\item{mcontrol}{control for missing data handling}

\item{current_block}{index of current block}

\item{pred}{predictions of current block}

\item{liste}{list with offsets}

\item{X}{original data}

\item{blocks}{information which variable belongs to which block}

\item{current_intercept}{the intercept estimated for the current block}
}
\value{
list with offsets, used imputation model and the blocks used for the imputation model (if applicable)
}
\description{
Calculates the offsets for the current block
}
