% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_boolean.R
\name{compare_boolean}
\alias{compare_boolean}
\title{Compare the rows of a matrix with a pattern}
\usage{
compare_boolean(object, pattern)
}
\arguments{
\item{object}{matrix}

\item{pattern}{pattern which is compared against the rows of the matrix}
}
\value{
logical vector if the pattern matches the rows
}
\description{
Compare the rows of a matrix with a pattern
}
