% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.control.R
\name{missing.control}
\alias{missing.control}
\title{Construct control structures for handling of missing data for \code{prioritylasso}}
\usage{
missing.control(
  handle.missingdata = c("none", "ignore", "impute.offset"),
  offset.firstblock = c("zero", "intercept"),
  impute.offset.cases = "complete.cases",
  nfolds.imputation = 10,
  lambda.imputation = c("lambda.min", "lambda.1se"),
  perc.comp.cases.warning = 0.3
)
}
\arguments{
\item{handle.missingdata}{how blockwise missing data should be treated. Default is \code{none} which does nothing, \code{ignore} ignores the observations with missing data for the current block, \code{impute.offset} imputes the offset for the missing values.}

\item{offset.firstblock}{determines if the offset of the first block for missing observations is zero or the intercept of the observed values for \code{handle.missingdata = ignore}}

\item{impute.offset.cases}{which cases should be used for the imputation model to impute missing offsets. So far, only complete cases are supported}

\item{nfolds.imputation}{nfolds for the glmnet of the imputation model}

\item{lambda.imputation}{which lambda-value should be used for predicting the imputed offsets in cv.glmnet}

\item{perc.comp.cases.warning}{percentage of complete cases when a warning is issued of too few cases for the imputation model}
}
\value{
list with control parameters
}
\description{
Construct control structures for handling of missing data for \code{prioritylasso}
}
