% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.prioritylasso.R
\name{coef.prioritylasso}
\alias{coef.prioritylasso}
\title{Extract coefficients from a prioritylasso object}
\usage{
\method{coef}{prioritylasso}(object, ...)
}
\arguments{
\item{object}{model of type prioritylasso}

\item{...}{additional arguments, currently not used}
}
\value{
List with the coefficients and the intercepts
}
\description{
Extract coefficients from a prioritylasso object
}
